KAS-1396: linked document model mixin  - gone
