+++ app/models/agendaitem.js
@@ -4,14 +4,13 @@ import {inject} from '@ember/service';
 import CONFIG from 'fe-redpencil/utils/config';
 import {alias} from '@ember/object/computed';
 import ModelWithModifier from 'fe-redpencil/models/model-with-modifier';
-import LinkedDocumentModelMixin from 'fe-redpencil/mixins/models/linked-document-model-mixin';
 import VRDocumentName, {compareFunction} from 'fe-redpencil/utils/vr-document-name';
 import {A} from '@ember/array';
 import {sortDocuments, getDocumentsLength} from 'fe-redpencil/utils/documents';
 
 let {attr, belongsTo, hasMany, PromiseArray, PromiseObject} = DS;
 
-export default ModelWithModifier.extend(LinkedDocumentModelMixin, {
+export default ModelWithModifier.extend({
   modelName: alias('constructor.modelName'),
   agendaService: inject(),
   addedAgendaitems: alias('agendaService.addedAgendaitems'),
@@ -52,14 +51,38 @@ export default ModelWithModifier.extend(LinkedDocumentModelMixin, {
     });
   }),
 
-
   documentsLength: computed('documents', function () {
-    return getDocumentsLength(this);
+    return getDocumentsLength(this, 'documents');
+  }),
+
+  linkedDocumentsLength: computed('linkedDocuments', function () {
+    return getDocumentsLength(this, 'linkedDocuments');
   }),
 
   documents: computed('documentVersions.@each.name', function () {
     return PromiseArray.create({
       promise: this.get('documentVersions').then((documentVersions) => {
+        if (documentVersions && documentVersions.get('length') > 0) {
+          const documentVersionIds = documentVersions.mapBy('id').join(',');
+          return this.store.query('document', {
+            filter: {
+              'documents': {id: documentVersionIds},
+            },
+            page: {
+              size: documentVersions.get('length'), // # documents will always be <= # document versions
+            },
+            include: 'type,documents,documents.access-level,documents.next-version,documents.previous-version',
+          }).then((containers) => {
+            return sortDocuments(this, containers);
+          });
+        }
+      })
+    });
+  }),
+
+  linkedDocuments: computed('linkedDocumentVersions.@each', function () {
+    return PromiseArray.create({
+      promise: this.get('linkedDocumentVersions').then((documentVersions) => {
         if (documentVersions && documentVersions.get('length') > 0) {
           const documentVersionIds = documentVersions.mapBy('id').join(',');
           return this.store.query('document', {
